Fix in DataUnit.
Issue with outgoing messages is gone.

+++ oscar/src/dataunit.h
@@ -179,7 +179,7 @@ struct toDataUnitHelper<T, true>
 	static inline QByteArray toByteArray(const QString &data, QTextCodec *codec = Util::defaultCodec(), ByteOrder bo = BigEndian)
 	{
 		QByteArray buf = toDataUnitHelper<QString>::toByteArray(data, codec);
-		if (static_cast<int>(std::numeric_limits<T>().max()) < buf.size())
+		if (std::numeric_limits<T>().max() < buf.size())
 			buf.resize(std::numeric_limits<T>().max() - 1);
 		return toByteArray(buf.size(), bo) + buf;
 	}
